requirements.txt
asgiref==3.8.1
botocore==1.38.9
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cloudinary==1.43.0
cryptography==44.0.0
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.25
django-allauth==0.50.0
django-cloudinary-storage==0.3.0
django-countries==7.2.1
django-crispy-forms==1.14.0
django-mathfilters==1.0.0
django-storages==1.14.4
gunicorn==23.0.0
idna==3.10
jmespath==1.0.1
oauthlib==3.2.2
packaging==24.2
pillow==10.3.0
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
s3transfer==0.11.2
setuptools==75.8.0
six==1.17.0
sqlparse==0.5.3
stripe==11.5.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
whitenoise==6.9.0
